deal slice write

+++ src/common/fs_types.h
@@ -33,7 +33,7 @@
 typedef struct fs_block_key {
     int64_t inode;
     int64_t offset;   //aligned by block size
-    uint32_t hash_code;
+    uint64_t hash_code;
 } FSBlockKey;
 
 typedef struct {
